fix(firmware/avatar): replace %zu with %u in ESP_LOG to avoid nano-printf crash
ESP-IDF defaults to newlib-nano printf (CONFIG_NEWLIB_NANO_FORMAT=y)
which does NOT support %zu. A stray %zu in a format string is silently
skipped, shifting downstream arguments by one — so a subsequent %s
ends up reading the size_t value as a const char* and dereferences it,
blowing up with LoadProhibited.

Symptom seen during Phase 4.5 end-to-end test: stack-chan rebooted
("ピロン" startup chime + LCD reset) immediately after AvatarSetFetcher
finished downloading the avatar set. The crash address was 0x83400 =
decimal 537,600, which is exactly kLayeredPayloadBytes — the size_t
argument that nano-printf shoved into the %s slot.

Fixes by casting size_t to unsigned int and using %u (32-bit-safe on
ESP32-S3) in all six log sites under avatar_set.cc /
avatar_set_fetcher.cc. The single-arg "Avatar set loaded: bytes=%zu"
log earlier did not crash because it had no downstream %s to misread;
only the fetcher's three-arg "...bytes=%zu, sha256=%s)" tipped over.

Debugging note: this only became diagnosable after rebuilding with
CONFIG_COMPILER_OPTIMIZATION_DEBUG=y. The release (-Os) backtrace had
addr2line pointing at AvatarSet::Unload (a completely unrelated
function) because aggressive inlining mapped the panic PC into
nearby code. Always use a debug build for panic post-mortem.

+++ firmware/main/boards/stackchan/avatar_set_fetcher.cc
@@ -62,8 +62,9 @@ void AvatarSetFetcher::Fetch(
 
     const size_t content_length = http->GetBodyLength();
     if (content_length != expected_size) {
-        ESP_LOGW(TAG, "Fetch: Content-Length mismatch (got=%zu, expected=%zu)",
-                 content_length, expected_size);
+        ESP_LOGW(TAG, "Fetch: Content-Length mismatch (got=%u, expected=%u)",
+                 static_cast<unsigned int>(content_length),
+                 static_cast<unsigned int>(expected_size));
         http->Close();
         on_complete(false, "", "content_length_mismatch");
         return;
@@ -81,7 +82,8 @@ void AvatarSetFetcher::Fetch(
     uint8_t* buffer = static_cast<uint8_t*>(
         heap_caps_malloc(expected_size, MALLOC_CAP_SPIRAM));
     if (buffer == nullptr) {
-        ESP_LOGE(TAG, "Fetch: PSRAM staging allocation failed (size=%zu)", expected_size);
+        ESP_LOGE(TAG, "Fetch: PSRAM staging allocation failed (size=%u)",
+                 static_cast<unsigned int>(expected_size));
         http->Close();
         on_complete(false, "", "psram_oom");
         return;
@@ -93,7 +95,8 @@ void AvatarSetFetcher::Fetch(
         int n = http->Read(reinterpret_cast<char*>(buffer + total_read),
                            static_cast<int>(to_read));
         if (n <= 0) {
-            ESP_LOGE(TAG, "Fetch: Read failed at offset %zu (n=%d)", total_read, n);
+            ESP_LOGE(TAG, "Fetch: Read failed at offset %u (n=%d)",
+                     static_cast<unsigned int>(total_read), n);
             heap_caps_free(buffer);
             http->Close();
             on_complete(false, "", "read_failed");
@@ -121,7 +124,14 @@ void AvatarSetFetcher::Fetch(
         return;
     }
 
-    ESP_LOGI(TAG, "Fetch: avatar set loaded (mode=%d, bytes=%zu, sha256=%s)",
-             static_cast<int>(mode), expected_size, actual_sha256.c_str());
+    // ESP-IDF defaults to newlib-nano printf which does NOT support %zu;
+    // a stray %zu is silently skipped, shifting downstream arguments — in
+    // particular the next %s reads a size_t value as a const char* and
+    // dereferences it, blowing up with LoadProhibited @ <that size>.
+    // Cast size_t to unsigned int and use %u to stay nano-printf-safe.
+    ESP_LOGI(TAG, "Fetch: avatar set loaded (mode=%d, bytes=%u, sha256=%s)",
+             static_cast<int>(mode),
+             static_cast<unsigned int>(expected_size),
+             actual_sha256.c_str());
     on_complete(true, actual_sha256, "");
 }
